# internal/api/server/api_operator.go
// SPDX-FileCopyrightText: Ella Networks Inc.
// SPDX-License-Identifier: BUSL-1.1

package server

import (
	"encoding/hex"
	"encoding/json"
	"fmt"
	"net/http"
	"net/netip"
	"strconv"
	"strings"

	"github.com/ellanetworks/core/internal/db"
	"github.com/ellanetworks/core/internal/logger"
	"go.uber.org/zap"
)

const (
	MaxSupportedTACs = 12
)

type UpdateOperatorTrackingParams struct {
	SupportedTacs []string `json:"supportedTacs"`
}

type UpdateOperatorIDParams struct {
	Mcc string `json:"mcc"`
	Mnc string `json:"mnc"`
}

type UpdateOperatorCodeParams struct {
	OperatorCode string `json:"operatorCode"`
}

type UpdateOperatorNASSecurityParams struct {
	Ciphering []string `json:"ciphering"`
	Integrity []string `json:"integrity"`
}

type UpdateOperatorSPNParams struct {
	FullName  string `json:"fullName"`
	ShortName string `json:"shortName"`
}

type UpdateOperatorSMSParams struct {
	SMSNumber string `json:"smsNumber"`
}

type GetOperatorSMSResponse struct {
	SMSNumber string `json:"smsNumber"`
}

type UpdateOperatorVoiceParams struct {
	PCSCFAddresses []string `json:"pcscfAddresses"`
}

type GetOperatorVoiceResponse struct {
	PCSCFAddresses []string `json:"pcscfAddresses"`
}

type GetOperatorTrackingResponse struct {
	SupportedTacs []string `json:"supportedTacs"`
}

type GetOperatorSPNResponse struct {
	FullName  string `json:"fullName"`
	ShortName string `json:"shortName"`
}

type GetOperatorResponse struct {
	ID              GetOperatorIDResponse          `json:"id,omitempty"`
	Tracking        GetOperatorTrackingResponse    `json:"tracking,omitempty"`
	HomeNetworkKeys []HomeNetworkKeyResponse       `json:"homeNetworkKeys"`
	NASSecurity     GetOperatorNASSecurityResponse `json:"nasSecurity"`
	SPN             GetOperatorSPNResponse         `json:"spn"`
	SMS             GetOperatorSMSResponse         `json:"sms"`
	Voice           GetOperatorVoiceResponse       `json:"voice"`
}

type GetOperatorIDResponse struct {
	Mcc string `json:"mcc"`
	Mnc string `json:"mnc"`
}

type GetOperatorNASSecurityResponse struct {
	Ciphering []string `json:"ciphering"`
	Integrity []string `json:"integrity"`
}

const (
	UpdateOperatorTrackingAction    = "update_operator_tracking"
	UpdateOperatorIDAction          = "update_operator_id"
	UpdateOperatorCodeAction        = "update_operator_code"
	UpdateOperatorNASSecurityAction = "update_operator_nas_security"
	UpdateOperatorSPNAction         = "update_operator_spn"
	UpdateOperatorSMSAction         = "update_operator_sms"
	UpdateOperatorVoiceAction       = "update_operator_voice"
)

func isValidMcc(mcc string) bool {
	if len(mcc) != 3 {
		return false
	}

	for _, c := range mcc {
		if c < '0' || c > '9' {
			return false
		}
	}

	return true
}

func isValidMnc(mnc string) bool {
	if len(mnc) != 2 && len(mnc) != 3 {
		return false
	}

	for _, c := range mnc {
		if c < '0' || c > '9' {
			return false
		}
	}

	return true
}

func isValidOperatorCode(operatorCode string) bool {
	if len(operatorCode) != 32 {
		logger.APILog.Warn("Invalid operator code length", zap.Int("length", len(operatorCode)))
		return false
	}

	_, err := hex.DecodeString(operatorCode)
	if err != nil {
		logger.APILog.Warn("Invalid operator code: not valid hex", zap.Error(err))
		return false
	}

	return true
}

func isValidTac(s string) bool {
	if len(s) != 6 {
		return false
	}

	if _, err := hex.DecodeString(s); err != nil {
		return false
	}

	return true
}

func isValidSst(sst int) bool {
	return sst >= 0 && sst <= 0xFF
}

func ParseSDString(s string) ([]byte, error) {
	if s == "" {
		return nil, nil
	}

	if strings.HasPrefix(s, "0x") || strings.HasPrefix(s, "0X") {
		return nil, fmt.Errorf("SD must not start with 0x")
	}

	b, err := hex.DecodeString(s)
	if err != nil {
		return nil, fmt.Errorf("invalid hex string: %w", err)
	}

	if len(b) > 3 {
		return nil, fmt.Errorf("SD must be at most 3 bytes (6 hex characters)")
	}

	arr := make([]byte, 3)
	copy(arr, b)

	return arr, nil
}

func GetOperator(dbInstance *db.Database) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		dbOperator, err := dbInstance.GetOperator(r.Context())
		if err != nil {
			writeError(r.Context(), w, http.StatusNotFound, "Operator not found", err, logger.APILog)
			return
		}

		hnKeys, err := dbInstance.ListHomeNetworkKeys(r.Context())
		if err != nil {
			logger.APILog.Warn("Failed to list home network keys", zap.Error(err))
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to list home network keys", err, logger.APILog)

			return
		}

		keyResponses := make([]HomeNetworkKeyResponse, 0, len(hnKeys))
		for _, k := range hnKeys {
			pubKey, err := k.GetPublicKey()
			if err != nil {
				logger.APILog.Warn("Failed to derive public key", zap.String("id", k.ID), zap.Error(err))
				writeError(r.Context(), w, http.StatusInternalServerError, "Failed to derive public key", err, logger.APILog)

				return
			}

			keyResponses = append(keyResponses, HomeNetworkKeyResponse{
				ID:            k.ID,
				KeyIdentifier: k.KeyIdentifier,
				Scheme:        k.Scheme,
				PublicKey:     pubKey,
			})
		}

		supportedTACs, err := dbOperator.GetSupportedTacs()
		if err != nil {
			logger.APILog.Warn("Failed to get supported TACs", zap.Error(err))
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to get supported TACs", err, logger.APILog)

			return
		}

		cipheringOrder, err := dbOperator.GetCiphering()
		if err != nil {
			logger.APILog.Warn("Failed to get ciphering order", zap.Error(err))
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to get ciphering order", err, logger.APILog)

			return
		}

		integrityOrder, err := dbOperator.GetIntegrity()
		if err != nil {
			logger.APILog.Warn("Failed to get integrity order", zap.Error(err))
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to get integrity order", err, logger.APILog)

			return
		}

		smsSettings, err := dbInstance.GetSMSSettings(r.Context())
		if err != nil {
			logger.APILog.Warn("Failed to get SMS settings", zap.Error(err))
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to get SMS settings", err, logger.APILog)

			return
		}

		pcscfAddresses, err := dbInstance.ListPCSCFAddresses(r.Context())
		if err != nil {
			logger.APILog.Warn("Failed to list P-CSCF addresses", zap.Error(err))
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to list P-CSCF addresses", err, logger.APILog)

			return
		}

		operator := &GetOperatorResponse{
			ID: GetOperatorIDResponse{
				Mcc: dbOperator.Mcc,
				Mnc: dbOperator.Mnc,
			},
			Tracking: GetOperatorTrackingResponse{
				SupportedTacs: supportedTACs,
			},
			HomeNetworkKeys: keyResponses,
			NASSecurity: GetOperatorNASSecurityResponse{
				Ciphering: cipheringOrder,
				Integrity: integrityOrder,
			},
			SPN: GetOperatorSPNResponse{
				FullName:  dbOperator.SpnFullName,
				ShortName: dbOperator.SpnShortName,
			},
			SMS: GetOperatorSMSResponse{
				SMSNumber: formatE164(smsSettings.SMSNumber),
			},
			Voice: GetOperatorVoiceResponse{
				PCSCFAddresses: addrStrings(pcscfAddresses),
			},
		}

		writeResponse(r.Context(), w, operator, http.StatusOK, logger.APILog)
	})
}

func UpdateOperatorTracking(dbInstance *db.Database) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		emailAny := r.Context().Value(contextKeyEmail)

		email, ok := emailAny.(string)
		if !ok {
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to get email", nil, logger.APILog)
			return
		}

		var params UpdateOperatorTrackingParams
		if err := json.NewDecoder(r.Body).Decode(&params); err != nil {
			writeError(r.Context(), w, http.StatusBadRequest, "Invalid request data", err, logger.APILog)
			return
		}

		if len(params.SupportedTacs) == 0 {
			writeError(r.Context(), w, http.StatusBadRequest, "supportedTacs is missing", nil, logger.APILog)
			return
		}

		if len(params.SupportedTacs) > MaxSupportedTACs {
			writeError(r.Context(), w, http.StatusBadRequest, "Too many supported TACs. Maximum is "+strconv.Itoa(MaxSupportedTACs), nil, logger.APILog)
			return
		}

		for _, tac := range params.SupportedTacs {
			if !isValidTac(tac) {
				writeError(r.Context(), w, http.StatusBadRequest, "Invalid TAC format. Must be a 3 bytes hex string", nil, logger.APILog)
				return
			}
		}

		if err := dbInstance.UpdateOperatorTracking(r.Context(), params.SupportedTacs); err != nil {
			logger.APILog.Warn("Failed to update operator tracking information", zap.Error(err))
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to update operator tracking information", err, logger.APILog)

			return
		}

		resp := SuccessResponse{Message: "Operator tracking information updated successfully"}
		writeResponse(r.Context(), w, resp, http.StatusCreated, logger.APILog)

		logger.LogAuditEvent(
			r.Context(),
			UpdateOperatorTrackingAction,
			email,
			getClientIP(r),
			"User updated operator tracking information",
		)
	})
}

func UpdateOperatorID(dbInstance *db.Database) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		emailAny := r.Context().Value(contextKeyEmail)

		email, ok := emailAny.(string)
		if !ok {
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to get email", nil, logger.APILog)
			return
		}

		var params UpdateOperatorIDParams
		if err := json.NewDecoder(r.Body).Decode(&params); err != nil {
			writeError(r.Context(), w, http.StatusBadRequest, "Invalid request data", err, logger.APILog)
			return
		}

		if params.Mcc == "" {
			writeError(r.Context(), w, http.StatusBadRequest, "mcc is missing", nil, logger.APILog)
			return
		}

		if params.Mnc == "" {
			writeError(r.Context(), w, http.StatusBadRequest, "mnc is missing", nil, logger.APILog)
			return
		}

		if !isValidMcc(params.Mcc) {
			writeError(r.Context(), w, http.StatusBadRequest, "Invalid mcc format. Must be a 3-decimal digit.", nil, logger.APILog)
			return
		}

		if !isValidMnc(params.Mnc) {
			writeError(r.Context(), w, http.StatusBadRequest, "Invalid mnc format. Must be a 2 or 3-decimal digit.", nil, logger.APILog)
			return
		}

		numSubs, err := dbInstance.CountSubscribers(r.Context())
		if err != nil {
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to count subscribers", err, logger.APILog)
			return
		}

		if numSubs > 0 {
			writeError(r.Context(), w, http.StatusBadRequest, "Cannot update operator ID when there are subscribers", nil, logger.APILog)
			return
		}

		if err := dbInstance.UpdateOperatorID(r.Context(), params.Mcc, params.Mnc); err != nil {
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to update operatorID", err, logger.APILog)
			return
		}

		resp := SuccessResponse{Message: "Operator ID updated successfully"}
		writeResponse(r.Context(), w, resp, http.StatusCreated, logger.APILog)

		logger.LogAuditEvent(
			r.Context(),
			UpdateOperatorIDAction,
			email,
			getClientIP(r),
			"User updated operator with Id: "+params.Mcc+params.Mnc,
		)
	})
}

func UpdateOperatorCode(dbInstance *db.Database) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		emailAny := r.Context().Value(contextKeyEmail)

		email, ok := emailAny.(string)
		if !ok {
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to get email", nil, logger.APILog)
			return
		}

		var params UpdateOperatorCodeParams
		if err := json.NewDecoder(r.Body).Decode(&params); err != nil {
			writeError(r.Context(), w, http.StatusBadRequest, "Invalid request data", err, logger.APILog)
			return
		}

		if params.OperatorCode == "" {
			writeError(r.Context(), w, http.StatusBadRequest, "operator code is missing", nil, logger.APILog)
			return
		}

		if !isValidOperatorCode(params.OperatorCode) {
			writeError(r.Context(), w, http.StatusBadRequest, "Invalid operator code format. Must be a 32-character hexadecimal string.", nil, logger.APILog)
			return
		}

		numSubs, err := dbInstance.CountSubscribers(r.Context())
		if err != nil {
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to count subscribers", err, logger.APILog)
			return
		}

		if numSubs > 0 {
			writeError(r.Context(), w, http.StatusBadRequest, "Cannot update operator code when there are subscribers", nil, logger.APILog)
			return
		}

		if err := dbInstance.UpdateOperatorCode(r.Context(), params.OperatorCode); err != nil {
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to update operatorID", err, logger.APILog)
			return
		}

		writeResponse(r.Context(), w, SuccessResponse{Message: "Operator Code updated successfully"}, http.StatusCreated, logger.APILog)

		logger.LogAuditEvent(
			r.Context(),
			UpdateOperatorCodeAction,
			email,
			getClientIP(r),
			"User updated operator Code",
		)
	})
}

var validCipheringAlgorithms = map[string]bool{"NULL": true, "SNOW3G": true, "AES": true}

var validIntegrityAlgorithms = map[string]bool{"NULL": true, "SNOW3G": true, "AES": true}

func isValidAlgorithmOrder(order []string, valid map[string]bool) (string, bool) {
	if len(order) == 0 {
		return "", false
	}

	if len(order) > 3 {
		return "", false
	}

	seen := make(map[string]bool, len(order))

	for _, alg := range order {
		if !valid[alg] {
			return alg, false
		}

		if seen[alg] {
			return alg, false
		}

		seen[alg] = true
	}

	return "", true
}

func UpdateOperatorNASSecurity(dbInstance *db.Database) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		emailAny := r.Context().Value(contextKeyEmail)

		email, ok := emailAny.(string)
		if !ok {
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to get email", nil, logger.APILog)
			return
		}

		var params UpdateOperatorNASSecurityParams
		if err := json.NewDecoder(r.Body).Decode(&params); err != nil {
			writeError(r.Context(), w, http.StatusBadRequest, "Invalid request data", err, logger.APILog)
			return
		}

		if len(params.Ciphering) == 0 {
			writeError(r.Context(), w, http.StatusBadRequest, "ciphering is required and must not be empty", nil, logger.APILog)
			return
		}

		if len(params.Integrity) == 0 {
			writeError(r.Context(), w, http.StatusBadRequest, "integrity is required and must not be empty", nil, logger.APILog)
			return
		}

		if badAlg, valid := isValidAlgorithmOrder(params.Ciphering, validCipheringAlgorithms); !valid {
			if badAlg != "" {
				writeError(r.Context(), w, http.StatusBadRequest, fmt.Sprintf("Invalid or duplicate ciphering algorithm: %s. Allowed: NULL, SNOW3G, AES", badAlg), nil, logger.APILog)
			} else {
				writeError(r.Context(), w, http.StatusBadRequest, "Maximum 3 ciphering algorithms allowed", nil, logger.APILog)
			}

			return
		}

		if badAlg, valid := isValidAlgorithmOrder(params.Integrity, validIntegrityAlgorithms); !valid {
			if badAlg != "" {
				writeError(r.Context(), w, http.StatusBadRequest, fmt.Sprintf("Invalid or duplicate integrity algorithm: %s. Allowed: NULL, SNOW3G, AES", badAlg), nil, logger.APILog)
			} else {
				writeError(r.Context(), w, http.StatusBadRequest, "Maximum 3 integrity algorithms allowed", nil, logger.APILog)
			}

			return
		}

		if err := dbInstance.UpdateOperatorSecurityAlgorithms(r.Context(), params.Ciphering, params.Integrity); err != nil {
			logger.APILog.Warn("Failed to update operator security algorithms", zap.Error(err))
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to update operator security algorithms", err, logger.APILog)

			return
		}

		resp := SuccessResponse{Message: "Operator NAS security algorithms updated successfully"}
		writeResponse(r.Context(), w, resp, http.StatusCreated, logger.APILog)

		logger.LogAuditEvent(
			r.Context(),
			UpdateOperatorNASSecurityAction,
			email,
			getClientIP(r),
			"User updated operator NAS security algorithms",
		)
	})
}

const (
	maxSPNLength = 50
)

func UpdateOperatorSPN(dbInstance *db.Database) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		emailAny := r.Context().Value(contextKeyEmail)

		email, ok := emailAny.(string)
		if !ok {
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to get email", nil, logger.APILog)
			return
		}

		var params UpdateOperatorSPNParams
		if err := json.NewDecoder(r.Body).Decode(&params); err != nil {
			writeError(r.Context(), w, http.StatusBadRequest, "Invalid request data", err, logger.APILog)
			return
		}

		if params.FullName == "" {
			writeError(r.Context(), w, http.StatusBadRequest, "fullName is required and must not be empty", nil, logger.APILog)
			return
		}

		if params.ShortName == "" {
			writeError(r.Context(), w, http.StatusBadRequest, "shortName is required and must not be empty", nil, logger.APILog)
			return
		}

		if len(params.FullName) > maxSPNLength {
			writeError(r.Context(), w, http.StatusBadRequest, fmt.Sprintf("fullName must be at most %d characters", maxSPNLength), nil, logger.APILog)
			return
		}

		if len(params.ShortName) > maxSPNLength {
			writeError(r.Context(), w, http.StatusBadRequest, fmt.Sprintf("shortName must be at most %d characters", maxSPNLength), nil, logger.APILog)
			return
		}

		if err := dbInstance.UpdateOperatorSPN(r.Context(), params.FullName, params.ShortName); err != nil {
			logger.APILog.Warn("Failed to update operator SPN", zap.Error(err))
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to update operator SPN", err, logger.APILog)

			return
		}

		resp := SuccessResponse{Message: "Operator SPN updated successfully"}
		writeResponse(r.Context(), w, resp, http.StatusCreated, logger.APILog)

		logger.LogAuditEvent(
			r.Context(),
			UpdateOperatorSPNAction,
			email,
			getClientIP(r),
			"User updated operator SPN",
		)
	})
}

func UpdateOperatorSMS(dbInstance *db.Database) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		emailAny := r.Context().Value(contextKeyEmail)

		email, ok := emailAny.(string)
		if !ok {
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to get email", nil, logger.APILog)
			return
		}

		var params UpdateOperatorSMSParams
		if err := json.NewDecoder(r.Body).Decode(&params); err != nil {
			writeError(r.Context(), w, http.StatusBadRequest, "Invalid request data", err, logger.APILog)
			return
		}

		settings, message := smsSettingsFromParams(params)
		if message != "" {
			writeError(r.Context(), w, http.StatusBadRequest, message, nil, logger.APILog)
			return
		}

		if err := settings.Validate(); err != nil {
			writeError(r.Context(), w, http.StatusBadRequest, "Invalid SMS settings", err, logger.APILog)
			return
		}

		if err := dbInstance.UpdateSMSSettings(r.Context(), &settings); err != nil {
			logger.APILog.Warn("Failed to update operator SMS settings", zap.Error(err))
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to update operator SMS settings", err, logger.APILog)

			return
		}

		resp := SuccessResponse{Message: "Operator SMS settings updated successfully"}
		writeResponse(r.Context(), w, resp, http.StatusCreated, logger.APILog)

		detail := fmt.Sprintf("User updated operator SMS settings (SMS number %s)", formatE164(settings.SMSNumber))

		logger.LogAuditEvent(r.Context(), UpdateOperatorSMSAction, email, getClientIP(r), detail)
	})
}

func smsSettingsFromParams(params UpdateOperatorSMSParams) (db.SMSSettings, string) {
	number, ok := parseE164(params.SMSNumber)
	if !ok {
		return db.SMSSettings{}, "smsNumber must be an E.164 number: + followed by 1 to 15 digits, for example +15550001111"
	}

	return db.SMSSettings{SMSNumber: number}, ""
}

func UpdateOperatorVoice(dbInstance *db.Database) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		emailAny := r.Context().Value(contextKeyEmail)

		email, ok := emailAny.(string)
		if !ok {
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to get email", nil, logger.APILog)
			return
		}

		var params UpdateOperatorVoiceParams
		if err := json.NewDecoder(r.Body).Decode(&params); err != nil {
			writeError(r.Context(), w, http.StatusBadRequest, "Invalid request data", err, logger.APILog)
			return
		}

		addresses := make([]netip.Addr, 0, len(params.PCSCFAddresses))

		for _, s := range params.PCSCFAddresses {
			addr, err := netip.ParseAddr(strings.TrimSpace(s))
			if err != nil {
				writeError(r.Context(), w, http.StatusBadRequest, fmt.Sprintf("pcscfAddresses: %q is not an IPv4 or IPv6 address", s), nil, logger.APILog)
				return
			}

			addresses = append(addresses, addr)
		}

		if err := db.ValidatePCSCFAddresses(addresses); err != nil {
			writeError(r.Context(), w, http.StatusBadRequest, "Invalid P-CSCF addresses", err, logger.APILog)
			return
		}

		if err := dbInstance.ReplacePCSCFAddresses(r.Context(), addresses); err != nil {
			logger.APILog.Warn("Failed to update operator voice settings", zap.Error(err))
			writeError(r.Context(), w, http.StatusInternalServerError, "Failed to update operator voice settings", err, logger.APILog)

			return
		}

		resp := SuccessResponse{Message: "Operator voice settings updated successfully"}
		writeResponse(r.Context(), w, resp, http.StatusCreated, logger.APILog)

		pcscf := "none"
		if len(addresses) > 0 {
			pcscf = strings.Join(addrStrings(addresses), ", ")
		}

		detail := fmt.Sprintf("User updated operator voice settings (P-CSCF addresses %s)", pcscf)

		logger.LogAuditEvent(r.Context(), UpdateOperatorVoiceAction, email, getClientIP(r), detail)
	})
}

func addrStrings(addresses []netip.Addr) []string {
	out := make([]string, 0, len(addresses))
	for _, a := range addresses {
		out = append(out, a.String())
	}

	return out
}