# internal/api/server/openapi.yaml
openapi: "3.1.0"
info:
  title: Ella Core API
  description: |
    Ella Core REST API.

    ## Authentication

    `Authorization: Bearer <token>`, where the token is a JWT or an API token (prefix `ellacore_`).

    ## Response Envelope

    - Success: `{"result": <payload>}`
    - Error: `{"error": "message"}`

    ## Pagination

    List endpoints accept `page` (default 1) and `per_page` (default 25, max 100).
    Responses include `items`, `page`, `per_page`, and `total_count`.
  version: "1.0.0"
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: "{scheme}://{host}"
    variables:
      scheme:
        default: https
        enum: [https, http]
      host:
        default: localhost:5000
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/operator:
    get:
      operationId: getOperator
      tags: [Operator]
      summary: Get operator configuration
      description: Returns the operator configuration.
      responses:
        "200":
          description: Operator configuration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatorResponseEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/operator/tracking:
    get:
      operationId: getOperatorTracking
      tags: [Operator]
      summary: Get tracking areas
      deprecated: true
      description: Deprecated. Use `GET /api/v1/operator`.
      responses:
        "200":
          description: Tracking area configuration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatorTrackingResponseEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateOperatorTracking
      tags: [Operator]
      summary: Update tracking areas
      description: Sets the supported Tracking Area Codes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOperatorTrackingParams"
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/operator/id:
    get:
      operationId: getOperatorId
      tags: [Operator]
      summary: Get PLMN ID
      deprecated: true
      description: Deprecated. Use `GET /api/v1/operator`.
      responses:
        "200":
          description: PLMN ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatorIDResponseEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateOperatorId
      tags: [Operator]
      summary: Update PLMN ID
      description: Sets the MCC and MNC. Cannot be changed while subscribers exist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOperatorIDParams"
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/operator/code:
    put:
      operationId: updateOperatorCode
      tags: [Operator]
      summary: Update operator code
      description: Sets the operator code (OP). Cannot be changed while subscribers exist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOperatorCodeParams"
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/operator/nas-security:
    put:
      operationId: updateOperatorNASSecurity
      tags: [Operator]
      summary: Update NAS security algorithms
      description: Sets the NAS ciphering and integrity algorithm preference order.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOperatorNASSecurityParams"
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/operator/spn:
    put:
      operationId: updateOperatorSPN
      tags: [Operator]
      summary: Update Service Provider Name (SPN)
      description: Sets the full and short network names.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOperatorSPNParams"
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/operator/sms:
    put:
      operationId: updateOperatorSMS
      tags: [Operator]
      summary: Update SMS settings
      description: Sets Ella Core's SMS number.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOperatorSMSParams"
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/operator/voice:
    put:
      operationId: updateOperatorVoice
      tags: [Operator]
      summary: Update voice settings
      description: Sets the P-CSCF addresses of the IMS that Ella Core gives to UEs on the ims data network. The list is ordered by preference. UEs receive the addresses of the IP version they request. An empty list clears it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOperatorVoiceParams"
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # -- Data Networks -------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "JWT token from login or API token (prefixed `ellacore_`)."
...
  responses:
    Success:
      description: Operation succeeded.
      content:
        application/json:
          schema:
            type: object
            properties:
              result:
                $ref: "#/components/schemas/SuccessResponse"
    BadRequest:
      description: Invalid request parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    OperatorID:
      type: object
      properties:
        mcc:
          type: string
          description: "Mobile Country Code (3 digits)."
        mnc:
          type: string
          description: "Mobile Network Code (2-3 digits)."
      required: [mcc, mnc]
    OperatorTracking:
      type: object
      properties:
        supportedTacs:
          type: array
          items:
            type: string
          description: "List of 6-character hex Tracking Area Codes."
      required: [supportedTacs]
    OperatorResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/OperatorID"
        tracking:
          $ref: "#/components/schemas/OperatorTracking"
        homeNetworkKeys:
          type: array
          items:
            $ref: "#/components/schemas/HomeNetworkKeyResponse"
        nasSecurity:
          $ref: "#/components/schemas/OperatorNASSecurity"
        spn:
          $ref: "#/components/schemas/OperatorSPN"
        sms:
          $ref: "#/components/schemas/OperatorSMS"
        voice:
          $ref: "#/components/schemas/OperatorVoice"
    OperatorResponseEnvelope:
      type: object
      properties:
        result:
          $ref: "#/components/schemas/OperatorResponse"
    OperatorTrackingResponseEnvelope:
      type: object
      properties:
        result:
          $ref: "#/components/schemas/OperatorTracking"
    OperatorIDResponseEnvelope:
      type: object
      properties:
        result:
          $ref: "#/components/schemas/OperatorID"
    UpdateOperatorTrackingParams:
      type: object
      required: [supportedTacs]
      properties:
        supportedTacs:
          type: array
          maxItems: 12
          items:
            type: string
            pattern: "^[0-9a-fA-F]{6}$"
    UpdateOperatorIDParams:
      type: object
      required: [mcc, mnc]
      properties:
        mcc:
          type: string
          pattern: "^[0-9]{3}$"
          description: "3-digit Mobile Country Code."
        mnc:
          type: string
          pattern: "^[0-9]{2,3}$"
          description: "2 or 3-digit Mobile Network Code."
    UpdateOperatorCodeParams:
      type: object
      required: [operatorCode]
      properties:
        operatorCode:
          type: string
          pattern: "^[0-9a-fA-F]{32}$"
          description: "32-character hex operator code (OP)."
    HomeNetworkKeyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Key ID.
        keyIdentifier:
          type: integer
          minimum: 0
          maximum: 255
          description: Home Network Public Key Identifier.
        scheme:
          type: string
          enum: [A, B]
          description: "`A` (X25519) or `B` (P-256)."
        publicKey:
          type: string
          description: Public key (hex).
...
    UpdateOperatorNASSecurityParams:
      type: object
      required: [ciphering, integrity]
      properties:
        ciphering:
          type: array
          minItems: 1
          maxItems: 3
          uniqueItems: true
          items:
            type: string
            enum: [NULL, SNOW3G, AES]
          description: "Ciphering algorithms, in preference order."
        integrity:
          type: array
          minItems: 1
          maxItems: 3
          uniqueItems: true
          items:
            type: string
            enum: [NULL, SNOW3G, AES]
          description: "Integrity algorithms, in preference order."
    OperatorNASSecurity:
      type: object
      properties:
        ciphering:
          type: array
          items:
            type: string
            enum: [NULL, SNOW3G, AES]
          description: "Ciphering algorithms, in preference order."
        integrity:
          type: array
          items:
            type: string
            enum: [NULL, SNOW3G, AES]
          description: "Integrity algorithms, in preference order."
    OperatorSPN:
      type: object
      properties:
        fullName:
          type: string
          description: "Full network name."
        shortName:
          type: string
          description: "Short network name."
    OperatorSMS:
      type: object
      required: [smsNumber]
      properties:
        smsNumber:
          type: string
          description: "Ella Core's E.164 number."
...
    UpdateOperatorSMSParams:
      type: object
      required: [smsNumber]
      properties:
        smsNumber:
          type: string
          pattern: "^(\\+[1-9][0-9]{0,14})?$"
          description: "Ella Core's E.164 number."
    OperatorVoice:
      type: object
      required: [pcscfAddresses]
      properties:
        pcscfAddresses:
          type: array
          items:
            type: string
          description: "P-CSCF addresses of the IMS given to UEs, in preference order. Empty when unset."
    UpdateOperatorVoiceParams:
      type: object
      required: [pcscfAddresses]
      properties:
        pcscfAddresses:
          type: array
          maxItems: 6
          items:
            type: string
          description: "Unicast IPv4 and IPv6 addresses of the IMS, in preference order. At most 3 of each IP version, without duplicates."
    UpdateOperatorSPNParams:
      type: object
      required: [fullName, shortName]
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 50
          description: "Full network name (printable ASCII)."
        shortName:
          type: string
          minLength: 1
          maxLength: 50
          description: "Short network name (printable ASCII)."

    # -- Data Networks ---------------------------------------------------
...